Rename NRGConsole to TPPWHolesale
