+++ src/TPPWholesale/Data/Configuration.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Upmind\ProvisionProviders\DomainNames\NRGConsole\Data;
+namespace Upmind\ProvisionProviders\DomainNames\TPPWholesale\Data;
 
 use Upmind\ProvisionBase\Provider\DataSet\DataSet;
 use Upmind\ProvisionBase\Provider\DataSet\Rules;
 
 /**
- * NRGConsole configuration.
+ * TPPWholesale configuration.
  *
  * @property-read string $accountNo Account NO
  * @property-read string $userId User ID
  * @property-read string $password Password
- * @property-read bool|null $sandbox Make API requests against the sandbox environment
+ * @property-read string|null $api_hostname Optionally override API hostname
  */
 class Configuration extends DataSet
 {
@@ -23,7 +23,7 @@ class Configuration extends DataSet
             'accountNo' => ['required', 'string', 'min:6'],
             'userId' => ['required', 'string', 'min:6'],
             'password' => ['required', 'string', 'min:3'],
-            'sandbox' => ['nullable', 'boolean'],
+            'api_hostname' => ['domain_name'],
         ]);
     }
 }
